Add DataBase

+++ src/main/java/ru/yandex/practicum/model/Genre.java
@@ -17,6 +17,6 @@ public class Genre  implements Comparable<Genre> {
 
     @Override
     public int compareTo(Genre o) {
-        return (int) (this.getId() - o.getId());
+        return this.getId() - o.getId();
     }
 }
